backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]
pydantic
python-dotenv
pymongo
passlib